refactor(react/profile): :truck: Se cambia la nomenclatura para más consistencia
