+++ src/react/components/DasboardPanel/components/Profile/configs/update_name_toast.js
